Remove appgame folder
After looking into X3D more, it seems like the game code is ran as a separate "application", with everything being placed into the appgame folder to keep it separate from the new main function nnMain. It calls the game's main function through a function called AppMain, which points to this likely being the case. Asserts from XCX show they didn't arrange files like that in that game, so I think it's safe to say it wasn't the case in XC1 either. Regardless, it just makes things cleaner and also means I don't have to type out as long of header paths, so I was probably gonna do this eventually anyway :>